SpaceMouse: libspnavdev works with older USB devices.
Initialize `spndev_event ev = {};` so that older USB devices where the
translation and rotation are in separate HID reports work properly. Before
they caused "wild" movement rotation due to hals the members of the
structure being uninitialized.

Add the axis scaling factor to the AxisData structure.

Remap the axis (swap Y and Z and invert Y) so that the movement of the model
matches the movement of the "hat" on the 3Dconnexion device.